requirements.txt
google-cloud-storage
transformers
torch
scikit-learn
python-dotenv
accelerate
fastapi
uvicorn
httpx
openai>=1.0.0
tqdm>=4.65.0
google-generativeai
tiktoken
sentencepiece
